requirements_gpu.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.1.1
cycler==0.10.0
decorator==4.4.1
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.0.3
numpy==1.18.0
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.24.2
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
tensorboard==2.1.0
tensorflow-gpu==2.1.0
tensorflow-estimator==2.1.0
tensorflow-probability==0.8.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
moviepy
pytest
